+++ BuildSiteIndex/Program.cs
@@ -3,9 +3,9 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Globalization;
+using System.IO;
 using System.Linq;
 using System.Net.Http;
-using System.Net.Http.Formatting;
 using System.Net.Http.Headers;
 using System.Text;
 using System.Text.RegularExpressions;
@@ -16,9 +16,10 @@ using FileNaming;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 using Newtonsoft.Json.Serialization;
+using ObjectModel;
+using Scanner;
 using StorageHelpers;
-using Twaddle.Directory.Scanner;
-using Twaddle.Gallery.ObjectModel;
+using UploadData;
 
 namespace BuildSiteIndex
 {
@@ -763,67 +764,13 @@ namespace BuildSiteIndex
             var retry = 0;
             do
             {
-                uploaded = await UploadItem(itemToPost, progressText, item.UploadType);
+                uploaded = await UploadHelper.UploadItem(itemToPost, progressText, item.UploadType);
                 ++retry;
             } while (!uploaded && retry < maxRetries);
             return uploaded;
         }
 
-        private static async Task<bool> UploadItem(GallerySiteIndex itemToPost, string progressText,
-            UploadType uploadType)
-        {
-            try
-            {
-                using (var client = new HttpClient
-                {
-                    BaseAddress = new Uri(Settings.Default.WebServerBaseAddress),
-                    Timeout = TimeSpan.FromSeconds(200)
-                })
-                {
-                    Console.WriteLine("Uploading ({0}): {1}", MakeUploadTypeText(uploadType), progressText);
-
-                    client.DefaultRequestHeaders.Accept.Add(
-                        new MediaTypeWithQualityHeaderValue("application/json"));
-
-                    var formatter = new JsonMediaTypeFormatter
-                    {
-                        SerializerSettings = {ContractResolver = new DefaultContractResolver()},
-                        SupportedEncodings = {Encoding.UTF8},
-                        Indent = false
-                    };
-
-                    var response = await client.PostAsync("tasks/sync", itemToPost, formatter);
-                    Console.WriteLine("Status: {0}", response.StatusCode);
-
-                    if (response.IsSuccessStatusCode)
-                    {
-                        Console.WriteLine(await response.Content.ReadAsStringAsync());
-                        return true;
-                    }
-                    return false;
-                }
-            }
-            catch (Exception exception)
-            {
-                Console.WriteLine("Error: {0}", exception.Message);
-                return false;
-            }
-        }
-
-        private static string MakeUploadTypeText(UploadType uploadType)
-        {
-            switch (uploadType)
-            {
-                case UploadType.NewItem:
-                    return "New";
-
-                case UploadType.DeleteItem:
-                    return "Delete";
-
-                default:
-                    return "Existing";
-            }
-        }
+        
 
 
         private static GallerySiteIndex CreateItemToPost(UploadQueueItem item)
