+++ GallerySync.sln
@@ -20,6 +20,12 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StorageHelpers", "StorageHe
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DeleteImageIfCorrupt", "DeleteImageIfCorrupt\DeleteImageIfCorrupt.csproj", "{B4EC82DA-7E7A-41B2-A0E9-9311F91D3F9B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Images", "Images\Images.csproj", "{2425D0E9-202D-49CD-9363-138218D882CF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Upload", "Upload\Upload.csproj", "{BAD7ABE8-2341-4230-A36A-3EC8DE8B4F2E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UploadData", "UploadData\UploadData.csproj", "{74ED470B-4F10-40E7-9136-CC7F812B0820}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -54,6 +60,18 @@ Global
 		{B4EC82DA-7E7A-41B2-A0E9-9311F91D3F9B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B4EC82DA-7E7A-41B2-A0E9-9311F91D3F9B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B4EC82DA-7E7A-41B2-A0E9-9311F91D3F9B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2425D0E9-202D-49CD-9363-138218D882CF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2425D0E9-202D-49CD-9363-138218D882CF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2425D0E9-202D-49CD-9363-138218D882CF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2425D0E9-202D-49CD-9363-138218D882CF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BAD7ABE8-2341-4230-A36A-3EC8DE8B4F2E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BAD7ABE8-2341-4230-A36A-3EC8DE8B4F2E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BAD7ABE8-2341-4230-A36A-3EC8DE8B4F2E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BAD7ABE8-2341-4230-A36A-3EC8DE8B4F2E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{74ED470B-4F10-40E7-9136-CC7F812B0820}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{74ED470B-4F10-40E7-9136-CC7F812B0820}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{74ED470B-4F10-40E7-9136-CC7F812B0820}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{74ED470B-4F10-40E7-9136-CC7F812B0820}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
