+++ Images/ImageHelpers.cs
@@ -1,31 +1,22 @@
-﻿namespace OutputBuilderClient
+﻿using System;
+using System.Diagnostics.Contracts;
+using SixLabors.ImageSharp;
+using SixLabors.ImageSharp.PixelFormats;
+
+namespace Images
 {
-    using System;
-    using System.Diagnostics.Contracts;
-
-    using GraphicsMagick;
-
     internal static class ImageHelpers
     {
         public static bool IsValidJpegImage(byte[] bytes, string context)
         {
             try
             {
-                using (var image = new MagickImage())
+                using (Image.Load(bytes, out var format))
                 {
-                    image.Warning += (sender, e) =>
-                        {
-                            Console.WriteLine("Image Validate Error: {0}", context);
-                            Console.WriteLine("Image Validate Error: {0}", e.Message);
-                            throw e.Exception;
-                        };
-
-                    image.Read(bytes);
-
-                    return image.Format == MagickFormat.Jpeg || image.Format == MagickFormat.Jpg;
+                    return format.DefaultMimeType == "image/jpeg";
                 }
             }
-            catch (MagickException exception)
+            catch (Exception exception)
             {
                 Console.WriteLine("Error: {0}", context);
                 Console.WriteLine("Error: {0}", exception);
@@ -45,7 +36,8 @@
         /// <remarks>
         ///     Only 0, 90, 180 and 270 degrees are supported.
         /// </remarks>
-        public static void RotateImageIfNecessary(MagickImage image, int degrees)
+        public static void RotateImageIfNecessary<TPixel>(Image<TPixel> image, int degrees)
+            where TPixel : struct, IPixel<TPixel>
         {
             Contract.Requires(image != null);
 
@@ -57,15 +49,15 @@
                     return;
 
                 case 90: // Rotate 90 degrees clockwise
-                    image.Rotate(90);
+                    image.Mutate(ctx => ctx.Rotate(90));
                     return;
 
                 case 180: // Rotate upside down
-                    image.Rotate(180);
+                    image.Mutate(ctx => ctx.Rotate(180));
                     return;
 
                 case 270: // Rotate 90 degrees anti-clockwise
-                    image.Rotate(270);
+                    image.Mutate(ctx => ctx.Rotate(270));
                     return;
 
                 default: // unknown - so can't rotate;
