+++ OutputBuilderClient/OutputBuilderClient.csproj
@@ -14,4 +14,10 @@
     <AssemblyTitle>Gallery.OutputBuilderClient</AssemblyTitle>
     <Description>Photo Gallery Output Builder</Description>
   </PropertyGroup>
+    <ItemGroup>
+        <ProjectReference Include="..\FileNaming\FileNaming.csproj"/>
+        <ProjectReference Include="..\Images\Images.csproj"/>
+        <ProjectReference Include="..\ObjectModel\ObjectModel.csproj"/>
+        <ProjectReference Include="..\Scanner\Scanner.csproj"/>
+    </ItemGroup>
 </Project>
