+++ OutputBuilderClient/PhotoExtensionMethods.cs
@@ -1,9 +1,9 @@
-﻿using System.Linq;
+﻿using System.IO;
+using System.Linq;
 using System.Threading.Tasks;
-using Alphaleonis.Win32.Filesystem;
 using FileNaming;
+using ObjectModel;
 using OutputBuilderClient.Properties;
-using Twaddle.Gallery.ObjectModel;
 
 namespace OutputBuilderClient
 {
