+++ Scanner/FileEntry.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Twaddle.Directory.Scanner
+namespace Scanner
 {
     public class FileEntry
     {
