+++ UploadData/ItemUpdateHelpers.cs
@@ -2,9 +2,9 @@
 using System.Collections.Generic;
 using System.Linq;
 using FileNaming;
-using Twaddle.Gallery.ObjectModel;
+using ObjectModel;
 
-namespace BuildSiteIndex
+namespace UploadData
 {
     internal static class ItemUpdateHelpers
     {
